finalapiwithfireand motion/requirements.txt
autopep8==2.0.1
Babel==2.14.0
backcall==0.2.0
bcrypt==4.0.1
beautifulsoup4==4.12.2
billiard==4.2.0
bleach==6.1.0
CacheControl==0.14.0
cachetools==5.3.1
celery==5.3.6
certifi==2023.5.7
cffi==1.16.0
channels==4.0.0
charset-normalizer==3.2.0
clarifai==9.6.3
clarifai-grpc==9.6.1
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
comm==0.1.4
contextlib2==21.6.0
contourpy==1.2.0
cryptography==42.0.2
cvzone==1.6.1
cx-Freeze==6.15.13
cycler==0.12.1
dacite==1.8.1
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
dicom2nifti==2.4.9
distlib==0.3.7
Django==4.2.3
django-celery-results==2.5.1
djangorestframework==3.14.0
dlib==19.24.2
dm-tree==0.1.8
dnspython==2.3.0
email-validator==1.3.1
exceptiongroup==1.1.3
execnet==2.0.2
executing==1.2.0
face-recognition==1.3.0
face-recognition-models==0.3.0
fastjsonschema==2.19.0
filelock==3.12.2
firebase-admin==6.4.0
Flask-Bcrypt==1.0.1
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.2
Flask-WTF==1.1.1
flatbuffers==23.5.26
fonttools==4.47.0
fqdn==1.5.1
fsspec==2023.12.2
gast==0.4.0
glob2==0.7
google-api-core==2.17.0
google-api-python-client==2.117.0
google-auth==2.27.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.0.0
google-cloud-core==2.4.1
google-cloud-firestore==2.14.0
google-cloud-storage==2.14.0
google-crc32c==1.5.0
google-pasta==0.2.0
google-resumable-media==2.7.0
googleapis-common-protos==1.60.0
grpcio==1.60.1
grpcio-status==1.60.1
h5py==3.9.0
htmlmin==0.1.12
httplib2==0.22.0
idna==3.4
ImageHash==4.3.1
importlib-metadata==6.8.0
importlib-resources==6.1.1
iniconfig==2.0.0
ipykernel==6.25.0
ipython==8.14.0
isoduration==20.11.0
jedi==0.19.0
Jinja2==3.1.2
joblib==1.3.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
jupyter-events==0.9.0
jupyter-lsp==2.2.1
jupyter_client==8.3.0
jupyter_core==5.3.1
jupyter_server==2.12.1
jupyter_server_terminals==0.5.0
jupyterlab==4.0.9
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.2
keras==2.13.1
Keras-Preprocessing==1.1.2
kiwisolver==1.4.5
kombu==5.3.5
libclang==16.0.6
llvmlite==0.41.1
macholib==1.16.3
Markdown==3.4.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mediapipe==0.10.0
mistune==3.0.2
ml-dtypes==0.2.0
mock==5.1.0
monai==1.2.0
MouseInfo==0.1.3
mpmath==1.3.0
msgpack==1.0.7
multimethod==1.10
mysql-connector==2.2.9
namex==0.0.7
nbclient==0.9.0
nbconvert==7.13.1
nbformat==5.9.2
nest-asyncio==1.5.7
networkx==3.2.1
nibabel==5.1.0
notebook==7.0.6
notebook_shim==0.2.3
Nuitka==2.0
numba==0.58.1
numpy==1.24.3
oauthlib==3.2.2
opencv-contrib-python==4.9.0.80
opencv-python==4.9.0.80
opt-einsum==3.3.0
ordered-set==4.1.0
overrides==7.4.0
packaging==23.1
pandas==2.1.4
pandas-profiling==3.6.6
pandocfilters==1.5.0
parso==0.8.3
path==16.7.1
path.py==12.5.0
patsy==0.5.6
pexpect==4.8.0
phik==0.12.3
pickleshare==0.7.5
Pillow==9.4.0
pipenv==2023.12.1
platformdirs==3.9.1
pluggy==1.3.0
prometheus-client==0.19.0
prompt-toolkit==3.0.39
proto-plus==1.23.0
protobuf==4.25.3
psutil==5.9.5
psycopg==3.1.10
psycopg2-binary==2.9.7
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyAutoGUI==0.9.54
pycodestyle==2.10.0
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
pydicom==2.4.3
PyDirectInput==1.0.4
PyGetWindow==0.0.9
Pygments==2.15.1
pyinstaller==6.3.0
pyinstaller-hooks-contrib==2024.0
PyJWT==2.8.0
PyMsgBox==1.0.9
pynput==1.7.6
pyobjc-core==10.1
pyobjc-framework-ApplicationServices==10.1
pyobjc-framework-Cocoa==10.1
pyobjc-framework-CoreText==10.1
pyobjc-framework-Quartz==10.1
pyparsing==3.1.1
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.30
pytest==7.4.3
pytest-shutil==1.7.0
python-dateutil==2.8.2
python-gdcm==3.0.22
python-json-logger==2.0.7
python-rapidjson==1.10
pytweening==1.0.7
pytz==2023.3.post1
pywatchman==1.4.1
PyWavelets==1.5.0
pywifi==1.1.12
PyYAML==6.0.1
pyzmq==25.1.0
redis==5.0.1
referencing==0.32.0
requests==2.31.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.0
rpds-py==0.15.2
rsa==4.9
rubicon-objc==0.4.7
scikit-learn==1.2.2
scipy==1.11.1
seaborn==0.12.2
send-email==201212
Send2Trash==1.8.2
six==1.16.0
sklearn==0.0.post7
sniffio==1.3.0
sounddevice==0.4.6
soupsieve==2.5
SQLAlchemy==1.4.46
sqlparse==0.4.4
stack-data==0.6.2
statsmodels==0.14.1
sympy==1.12
tangled-up-in-unicode==0.2.0
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.36.0
tensorflow-macos==2.13.0
termcolor==2.3.0
terminado==0.18.0
threadpoolctl==3.2.0
tinycss2==1.2.1
tomli==2.0.1
torch==2.1.2
torchaudio==2.1.2
torchvision==0.16.2
tornado==6.3.2
tqdm==4.66.1
traitlets==5.9.0
tritonclient==2.34.0
typeguard==4.1.5
types-python-dateutil==2.8.19.14
typing_extensions==4.5.0
tzdata==2023.3
uri-template==1.3.0
uritemplate==4.1.1
urllib3==1.26.16
visions==0.7.5
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
websockets==12.0
Werkzeug==2.3.6
wordcloud==1.9.3
wrapt==1.14.1
WTForms==3.0.1
ydata-profiling==4.6.3
zipp==3.16.2
zstandard==0.22.0
